refactor(v2): split built-in source registry modules
