+++ tests/registry/test_catalog.py
@@ -19,7 +19,7 @@ from souwen.registry.catalog import (
     sources_by_category,
 )
 from souwen.registry.loader import lazy
-from souwen.registry import external_plugins, fetch_providers
+from souwen.registry import all_adapters, defaults_for, external_plugins, fetch_providers
 from souwen.registry.views import _reg_external
 from souwen.web.fetch import get_fetch_handlers
 
@@ -96,6 +96,129 @@ def test_default_source_map_references_existing_safe_entries() -> None:
             assert entry.visibility == "public"
             assert entry.risk_level != "high"
             assert entry.stability != "deprecated"
+
+
+def test_adapter_registration_order_snapshot() -> None:
+    assert tuple(all_adapters()) == (
+        "openalex",
+        "semantic_scholar",
+        "crossref",
+        "arxiv",
+        "dblp",
+        "core",
+        "pubmed",
+        "biorxiv",
+        "zotero",
+        "huggingface",
+        "europepmc",
+        "pmc",
+        "doaj",
+        "zenodo",
+        "hal",
+        "openaire",
+        "iacr",
+        "ieee_xplore",
+        "unpaywall",
+        "patentsview",
+        "pqai",
+        "epo_ops",
+        "uspto_odp",
+        "the_lens",
+        "cnipa",
+        "patsnap",
+        "google_patents",
+        "duckduckgo",
+        "duckduckgo_news",
+        "duckduckgo_images",
+        "duckduckgo_videos",
+        "yahoo",
+        "brave",
+        "google",
+        "bing",
+        "bing_cn",
+        "startpage",
+        "baidu",
+        "mojeek",
+        "yandex",
+        "serpapi",
+        "brave_api",
+        "serper",
+        "scrapingdog",
+        "metaso",
+        "tavily",
+        "exa",
+        "perplexity",
+        "firecrawl",
+        "linkup",
+        "xcrawl",
+        "zhipuai",
+        "aliyun_iqs",
+        "searxng",
+        "whoogle",
+        "websurfx",
+        "reddit",
+        "twitter",
+        "facebook",
+        "weibo",
+        "zhihu",
+        "youtube",
+        "bilibili",
+        "wikipedia",
+        "github",
+        "stackoverflow",
+        "csdn",
+        "juejin",
+        "linuxdo",
+        "nodeseek",
+        "hostloc",
+        "v2ex",
+        "coolapk",
+        "xiaohongshu",
+        "community_cn",
+        "feishu_drive",
+        "wayback",
+        "builtin",
+        "jina_reader",
+        "arxiv_fulltext",
+        "crawl4ai",
+        "scrapling",
+        "scrapfly",
+        "diffbot",
+        "scrapingbee",
+        "zenrows",
+        "scraperapi",
+        "apify",
+        "cloudflare",
+        "newspaper",
+        "readability",
+        "mcp",
+        "site_crawler",
+        "deepwiki",
+    )
+
+
+def test_default_source_order_snapshot() -> None:
+    assert {
+        "paper:search": defaults_for("paper", "search"),
+        "patent:search": defaults_for("patent", "search"),
+        "web:search": defaults_for("web", "search"),
+        "web:search_news": defaults_for("web", "search_news"),
+        "video:search": defaults_for("video", "search"),
+        "knowledge:search": defaults_for("knowledge", "search"),
+        "developer:search": defaults_for("developer", "search"),
+        "archive:archive_lookup": defaults_for("archive", "archive_lookup"),
+        "fetch:fetch": defaults_for("fetch", "fetch"),
+    } == {
+        "paper:search": ["openalex", "crossref", "arxiv", "dblp", "pubmed", "biorxiv"],
+        "patent:search": ["google_patents"],
+        "web:search": ["duckduckgo", "bing"],
+        "web:search_news": ["duckduckgo_news"],
+        "video:search": ["youtube", "bilibili"],
+        "knowledge:search": ["wikipedia"],
+        "developer:search": ["github", "stackoverflow"],
+        "archive:archive_lookup": ["wayback"],
+        "fetch:fetch": ["builtin"],
+    }
 
 
 def test_non_public_high_risk_and_deprecated_sources_are_not_default_available() -> None:
